requirements.txt
flask==3.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
markdown==3.5.1
pygments==2.17.2
bleach==6.1.0